Рефакторинг и упорядочивание файловой структуры
