+++ src/components/event/event-city-list.js
@@ -1,11 +1,11 @@
-const cityList = (cities) => {
+const generateCityList = (cities) => {
   return cities.map((city) => `<option value="${city}"></option>`).join(``);
 };
 
 export const createTripCityListTemplate = (cities) => {
   return `
     <datalist id="destination-list-1">
-        ${cityList(cities)}
+        ${generateCityList(cities)}
     </datalist>
   `;
 };
